+++ osgp-adapter-ws-distributionautomation/src/main/java/org/osgpfoundation/osgp/adapter/ws/da/application/services/ResendNotificationDistributionAutomation.java
@@ -1,15 +1,10 @@
 package org.osgpfoundation.osgp.adapter.ws.da.application.services;
 
-import java.util.ArrayList;
-import java.util.Date;
-import java.util.List;
-
 import org.apache.commons.lang3.EnumUtils;
 import org.osgpfoundation.osgp.adapter.ws.schema.distributionautomation.notification.NotificationType;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.scheduling.support.CronSequenceGenerator;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -19,16 +14,7 @@ import com.alliander.osgp.adapter.ws.shared.services.ResendNotificationService;
 
 @Service
 @Transactional(value = "transactionManager")
-public class ResendNotificationDistributionAutomation extends ResendNotificationService{
-
-	@Autowired
-	private String resendNotificationCronExpression;
-
-	@Autowired
-	private int resendNotificationMultiplier;
-
-	@Autowired
-	private int resendNotificationMaximum;
+public class ResendNotificationDistributionAutomation extends ResendNotificationService {
 
 	@Autowired
 	private NotificationServiceDistributionAutomation notificationServiceDistributionAutomation;
@@ -37,43 +23,19 @@ public class ResendNotificationDistributionAutomation extends ResendNotification
 	private ResponseDataRepository responseDataRepository;
 
 	private static final Logger LOGGER = LoggerFactory.getLogger(ResendNotificationDistributionAutomation.class);
-	
-	public void execute() {
 
-		List<ResponseData> notificationsToResend = new ArrayList<ResponseData>();
-		notificationsToResend = this.responseDataRepository
-				.findByNumberOfNotificationsSendLessThan(this.resendNotificationMaximum);
-		for (ResponseData responseData : notificationsToResend) {
-			if (EnumUtils.isValidEnum(NotificationType.class, responseData.getMessageType())) {
-				int multiplier = this.resendNotificationMultiplier;
-				if (responseData.getNumberOfNotificationsSend() == 0) {
-					multiplier = 1;
-				} else {
-					multiplier = responseData.getNumberOfNotificationsSend() * multiplier;
-				}
+	public void executer(ResponseData responseData) {
 
-				Date modificationDate = responseData.getModificationTime();
-				Date currentDate = new Date();
-				long previousModificationTime = currentDate.getTime() - modificationDate.getTime();
-
-				final CronSequenceGenerator cronSequenceGenerator = new CronSequenceGenerator(
-						this.resendNotificationCronExpression);
-				final Date nextFirstExecutionDate = cronSequenceGenerator.next(currentDate);
-				final Date nextSecondExecutionDate = cronSequenceGenerator.next(nextFirstExecutionDate);
-				final long cronSequenceInterval = nextSecondExecutionDate.getTime() - nextFirstExecutionDate.getTime();
-
-				if ((cronSequenceInterval * multiplier) < previousModificationTime) {
-					LOGGER.info("Found response data for resending notification");
-					NotificationType notificationType = NotificationType.valueOf(responseData.getMessageType());
-					this.notificationServiceDistributionAutomation.sendNotification(responseData.getOrganisationIdentification(),
-							responseData.getDeviceIdentification(), responseData.getResultType().name(),
-							responseData.getCorrelationUid(), getNotificationMessage(responseData.getMessageType()),
-							notificationType);
-					LOGGER.info("Notification has been resend");
-					responseData.setNumberOfNotificationsSend(responseData.getNumberOfNotificationsSend() + 1);
-					this.responseDataRepository.save(responseData);
-				}
-			}
+		if (EnumUtils.isValidEnum(NotificationType.class, responseData.getMessageType())) {
+			LOGGER.info("Found response data for resending notification");
+			NotificationType notificationType = NotificationType.valueOf(responseData.getMessageType());
+			this.notificationServiceDistributionAutomation.sendNotification(
+					responseData.getOrganisationIdentification(), responseData.getDeviceIdentification(),
+					responseData.getResultType().name(), responseData.getCorrelationUid(),
+					getNotificationMessage(responseData.getMessageType()), notificationType);
+			LOGGER.info("Notification has been resend");
+			responseData.setNumberOfNotificationsSend(responseData.getNumberOfNotificationsSend() + 1);
+			this.responseDataRepository.save(responseData);
 		}
 	}
 }
